LCARS button

An LCARS wrapper around shiny::actionButton with additional color control.

lcarsButton(
  inputId,
  label,
  icon = NULL,
  color = "atomic-tangerine",
  hover_color = "red-damask",
  ...
)

Arguments

inputId
The input slot that will be used to access the value.

label
The contents of the button or link–usually a text label, but you could also use any other HTML, like an image.

icon
An optional icon to appear on the button.

color
button color. Can be any color given in hex format. Named colors must be LCARS colors. See lcarsdata for options.

hover_color
Named colors must be LCARS colors. Other arbitrary colors are not supported for hovering. If hover_color = NULL, the button will darken on hover.

...
Named attributes to be applied to the button.

Value

HTML

See also

Examples

lcarsButton("btn", "A button")

#> <button class="btn btn-default action-button lcars-btn lcars-btn-red-damask" id="btn" style="background-color:#FF9900;" type="button">A button</button>