LCARS Shiny UI

Create a Shiny UI page with an LCARS theme.

lcarsPage(
  ...,
  title = NULL,
  force_uppercase = TRUE,
  label_uppercase = TRUE,
  lcars_font_headers = TRUE,
  lcars_font_labels = TRUE,
  lcars_font_text = TRUE
)

Arguments

...

The contents of the document body.

title

The browser window title (defaults to the host URL of the page).

force_uppercase

logical, see details.

label_uppercase

logical, also make widget labels uppercase globally.

lcars_font_headers

use LCARS-style font family for major heading text. See details.

lcars_font_labels

use LCARS-style font family for LCARS widget labels/titles.

lcars_font_text

use LCARS-style font family for general text (paragraphs, lists, etc.).

Value

A UI definition that can be passed to the shinyUI function.

Details

The LCARS style heavily emphasizes uppercase text. Set force_uppercase = TRUE to force this standard via CSS. This does not make everything uppercase; things like input labels are left alone (use label_uppercase = TRUE). However, text in general in uppercased.

Set these to FALSE if you need control over casing. This allows sensible judgment over how to balance the tension between making something that conforms well to the familiar LCARS aesthetic and making something that communicates information with a lower cognitive load for the user. Similarly, set lcars_font* arguments to FALSE to use a more readable sans serif font as desired. See examples for recommendations.

Examples

# \donttest{
# Recommended settings
# (results stored in x and not printed due to length)

# for a more standard LCARS style: default settings.
x <- lcarsPage()

# for a more readable style: less uppercase, switch to sans font
x <- lcarsPage(force_uppercase = FALSE, label_uppercase = FALSE,
  lcars_font_labels = FALSE, lcars_font_text = FALSE)
# }