LCARS radio buttons

A minimal replacement for radioButtons that provides two additional color arguments for consistency with functions like lcarsCheckbox.

lcarsRadio(inputId, label, choices = NULL, selected = NULL,
  inline = FALSE, width = NULL, choiceNames = NULL,
  choiceValues = NULL, label_color = "#FFFFFF",
  choice_color = label_color)

Arguments

inputId
character, the input slot that will be used to access the value.

label
character, display label for the control, or NULL for no label.

choices
see shiny::radioButtons for details.

selected
The initially selected value (if not specified then defaults to the first value).

inline
If TRUE, render the choices inline (i.e. horizontally).

width
a valid CSS unit.

choiceNames
see shiny::radioButtons for details.

choiceValues
see shiny::radioButtons for details.

label_color
Label color. Can be any color given in hex format. Named colors must be LCARS colors. See lcarsColors for options.

choice_color
choice text color, as above.

Value

A set of radio buttons that can be added to a UI definition.

Examples

## Only run examples in interactive R sessions
if (interactive()) {

  ui <- lcarsPage(
    lcarsRadio("dist", "Distribution type:",
                 c("Normal" = "norm",
                   "Uniform" = "unif",
                   "Log-normal" = "lnorm",
                   "Exponential" = "exp"),
               label_color = "lilac",
               choice_color = "atomic-tangerine"
              ),
    plotOutput("distPlot")
  )

  server <- function(input, output) {
    output$distPlot <- renderPlot({
      dist <- switch(input$dist,
                     norm = rnorm,
                     unif = runif,
                     lnorm = rlnorm,
                     exp = rexp,
                     rnorm)

      hist(dist(500))
    })
  }

  shinyApp(ui, server)
}